_main file!
I found this amazing template on here (I'll link it in the README.md) and thought it would be a wonderful start for a beginner portfolio. I customized it to my lil brand colors and added personal touches too! I know it's not the best work, but I spent two weeks on this and I'm so proud that I know more about VSCode!